PWA-25 spostato il servizio con il richiamo ad amchart
Spostato i servizi getDryerHistoryProduction e getLineHistoryProduction
